auth-service/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic[email]
pydantic-settings
PyJWT
requests
pika
httpx